changed dyamic route from id to email
